get it to build on android
this commit is nothing but desperate and hacky changes, in order to get
the build further along. do not mind the b00b135

the libcurl.a and libssl.a files were built using
https://github.com/ibaoger/libcurl-android

the codegen interprets android 0xb00b135 as an unbindable function, aka
a function that doesnt exist on android. for some functions this is
actually needed (like getControllerConnected), however for others i just
disabled them because the geode version doesnt match the android symbol
and i couldnt be bothered to fix it. when i wasnt lazy i did fix it
which is why many random functions in GeometryDash.bro got changed,
because they were just wrong from the android signature

many other changes were made to the codegen as well, such as disabling
the address for constructors and dtors, the reason for which being that
you cant take the address of those in c++, so i dont even know how this
built on windows

the paths in Dirs.cpp are also very messed up, need to clean those up.
as a result of trying to get it to not crash i changed up
ghc::filesystem::create_directory with the geode utils counterpart,
which ended up not being necessary after making the save dir the actual
geode launcher dir

btw have i mentioned that this is meant to be used with
https://github.com/qimiko/geode-launcher ?

this does not include my local copy of tuliphook, which had also lots of
changes to get it to build, mostly with the cmake to get capstone to
shut up (god i hate capstone its so bad)

dont mind the build scripts, they are made for mat and mat only